[INJIWEB-1592] refactor - mv useApi related data to related file
